% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_objects.R
\name{build_dslogitsb_config}
\alias{build_dslogitsb_config}
\title{DSLogitSB config builder}
\usage{
build_dslogitsb_config(
  m_0 = 0,
  a_sig2_psi = 2,
  b_sig2_psi = 1,
  a_sig2_alpha = 2,
  b_sig2_alpha = 1,
  a_sig2_beta = 2,
  b_sig2_beta = 1,
  a_sig2_mu = 2,
  b_sig2_mu = 1
)
}
\arguments{
\item{m_0}{common substructure mean hyperparameter}

\item{a_sig2_psi, b_sig2_psi}{independent substructure variance hyperparameters}

\item{a_sig2_alpha, b_sig2_alpha}{common-from substructure variance hyperparameters}

\item{a_sig2_beta, b_sig2_beta}{common-to substructure variance hyperparameters}

\item{a_sig2_mu, b_sig2_mu}{common substructure variance hyperparameters}
}
\value{
list containing metadata and relevant parameter info
}
\description{
Dependent Sojourns Logit Stick-Breaking (DSLogitSB)
}
