% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_objects.R
\name{build_gammamix_config}
\alias{build_gammamix_config}
\title{GammaMixture config builder}
\usage{
build_gammamix_config(
  n_component,
  a_kappa = 3,
  a_gam_kappa = 1,
  b_gam_kappa = 1,
  a_lam = 2.1,
  a_gam_lam = 1,
  b_gam_lam = 1
)
}
\arguments{
\item{n_component}{number of mixture components}

\item{a_kappa}{gamma kernel shape prior}

\item{a_gam_kappa, b_gam_kappa}{gamma kernel shape hyperparameters}

\item{a_lam}{gamma kernel scale prior}

\item{a_gam_lam, b_gam_lam}{gamma kernel scale hyperparameters}
}
\value{
list containing metadata and relevant parameter info
}
\description{
GammaMixture config builder
}
