% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_objects.R
\name{build_hrp_config}
\alias{build_hrp_config}
\title{HRP config builder}
\usage{
build_hrp_config()
}
\value{
list containing metadata and relevant parameter info
}
\description{
Homogeneous Renewal Process (HRP)
}
