% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_objects.R
\name{build_llhazardmix_config}
\alias{build_llhazardmix_config}
\title{LogLogisticHazardMixture config builder}
\usage{
build_llhazardmix_config(
  n_component,
  R_factor = 0.99,
  a_sig2 = 2.1,
  b_sig2 = 1,
  m_beta = 0,
  s_beta = 1000
)
}
\arguments{
\item{n_component}{number of mixture components}

\item{R_factor}{scale basis range by a factor within \link{0,1}.}

\item{a_sig2, b_sig2}{Gaussian process marginal variance parameters}

\item{m_beta, s_beta}{Gaussian process mean function parameters}
}
\value{
list containing metadata and relevant parameter info
}
\description{
LogLogisticHazardMixture config builder
}
