% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_objects.R
\name{build_logitsb_config}
\alias{build_logitsb_config}
\title{LogitSB config builder}
\usage{
build_logitsb_config(m_mu = 0, s_mu = 1000, a_sig2 = 1, b_sig2 = 1)
}
\arguments{
\item{m_mu, s_mulogit}{stick-breaking mean hyperparameters}

\item{a_sig2, b_sig2}{logit stick-breaking variance hyperparameters}
}
\value{
list containing metadata and relevant parameter info
}
\description{
Logit Stick-Breaking (LogitSB)
}
