% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_objects.R
\name{build_mrp_config}
\alias{build_mrp_config}
\title{MRP config builder}
\usage{
build_mrp_config(a_p = 0.1)
}
\arguments{
\item{a_p}{transition probability hyperparameter}
}
\value{
list containing metadata and relevant parameter info
}
\description{
Markov Renewal Process (MRP)
}
