% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_objects.R
\name{build_uniformmix_config}
\alias{build_uniformmix_config}
\title{UniformMixture config builder}
\usage{
build_uniformmix_config(
  n_component,
  a_eta = 1,
  b_eta = 1,
  a_gam = 1,
  b_gam = 1
)
}
\arguments{
\item{n_component}{number of mixture components}

\item{a_eta, b_eta}{inverse gamma shape hyperparameters}

\item{a_gam, b_gam}{inverse gamma scale hyperparameters}
}
\value{
list containing metadata and relevant parameter info
}
\description{
UniformMixture config builder
}
