% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_objects.R
\name{build_weibull_config}
\alias{build_weibull_config}
\title{Weibull config builder}
\usage{
build_weibull_config(a_alpha = 1, b_alpha = 1, a_lam = 1, b_lam = 1)
}
\arguments{
\item{a_alpha, b_alpha}{weibull shape hyperparameters}

\item{a_lam, b_lam}{weibull scale hyperparameters}
}
\value{
list containing metadata and relevant parameter info
}
\description{
Weibull config builder
}
