% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{fit_renewal_model}
\alias{fit_renewal_model}
\title{Fit a Bayesian Renewal Model}
\usage{
fit_renewal_model(
  julia,
  process_config,
  density_config,
  stickbreaking_config,
  sojourns,
  max_time,
  states,
  initial_state = 1,
  n_iteration = 5000,
  n_burnin = 1000,
  n_thin = 1,
  n_density_eval = 100,
  n_kfunction_eval = 30,
  eval_factor = 1,
  save_hazard = FALSE,
  save_kfunction = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{julia}{Julia object returned from \code{build_julia_interface}}

\item{process_config}{builder output for renewal process type: HRP, MRP}

\item{density_config}{builder output for density model:
WeibullModel, GammaMixtureModel, UniformMixtureModel}

\item{stickbreaking_config}{builder output for stick-breaking model:
DPSB, LogitSB, DSLogitSB. Is only required for mixture density models}

\item{sojourns}{vector of sojourn times or inter-arrival times}

\item{max_time}{upper bound on the observation time window}

\item{states}{integer vector of states for MRP process models}

\item{initial_state}{integer value indicating initial state for MRP process models}

\item{n_iteration, n_burnin, n_thin}{MCMC sampling output variables}

\item{eval_factor}{multiplier fot upper bound of evaluation grid
(normally at max(sojourns))}

\item{save_hazard, save_kfunction}{indicate whether to save these functions.
K-function only applies to HRP process models}

\item{verbose}{indicates whether to display sampling progress messages}

\item{n_dendsity_eval, n_kfunction_eval}{grid size for function evaluation.
K-function only applies to HRP process models}
}
\description{
Given a certain configuration setup, passes relevant data to julia, runs the
appropriate MCMC algorithm, and returns a list of posterior samples.
}
