% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_objects.R
\name{build_dpsb_config}
\alias{build_dpsb_config}
\title{DPSB config builder}
\usage{
build_dpsb_config(a_alpha = 5, b_alpha = 1)
}
\arguments{
\item{a_alpha, b_alpha}{DP concentration hyperparameters}
}
\value{
list containing metadata and relevant parameter info
}
\description{
Dirichlet Process Stick-Breaking (DPSB)
}
