% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_objects.R
\name{build_gammamix_config}
\alias{build_gammamix_config}
\title{GammaMixture config builder}
\usage{
build_gammamix_config(
  n_component,
  a_kappa = 3,
  a_gamma_kappa = 1,
  b_gamma_kappa = 1,
  a_lambda = 2.1,
  a_gamma_lambda = 1,
  b_gamma_lambda = 1
)
}
\arguments{
\item{n_component}{number of mixture components}

\item{a_kappa}{gamma kernel shape prior}

\item{a_gamma_kappa, b_gamma_kappa}{gamma kernel shape hyperparameters}

\item{a_lambda}{gamma kernel scale prior}

\item{a_gamma_lambda, b_gamma_lambda}{gamma kernel scale hyperparameters}
}
\value{
list containing metadata and relevant parameter info
}
\description{
GammaMixture config builder
}
