% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_objects.R
\name{build_modrp_config}
\alias{build_modrp_config}
\title{ModRP config builder}
\usage{
build_modrp_config(n_component, Lambda, a_alpha = 3, b_alpha = 1)
}
\arguments{
\item{n_component}{number of intensity mixture components}

\item{Lambda}{total modulating intensity. Values near n are reasonable.}

\item{a_alpha, b_alpha}{prior parameter values for Dirichlet process alpha}
}
\value{
list containing metadata and relevant parameter info
}
\description{
Modulated Renewal Process (ModRP)
}
